Fix some compiler warnings and continue refactor

+++ src/test/torrent.cpp
@@ -21,7 +21,7 @@ static std::string random_string(size_t length) {
             "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
             "abcdefghijklmnopqrstuvwxyz";
         const size_t max_index = (sizeof(charset) - 1);
-        return charset[rand() % max_index];
+        return charset[static_cast<std::size_t>(rand()) % max_index];
     };
     std::string str(length, 0);
     std::generate_n(str.begin(), length, randchar);
